Add retries if loading credentials fails with a SdkClientException (#60)
If the library fails to load credentials with a SdkClientException, we retry. The retry policy is an exponential back off policy with full jitter. By default the base retry interval is 500 ms, the max retry interval is 2000 ms and the max number of retries is 3. The max retry interval can be set via the option awsMaxBackOffTimeMs and the max number of retries can be set via awsMaxRetries.